requirements.txt
pybotters>=1.7.0
hyperliquid-python-sdk>=0.18.0
discord-webhook
python-dotenv
numpy
aiohttp
eth-account